[FIX] read does not need to have a dtype specified

+++ tests/test_read_data.py
@@ -16,7 +16,7 @@ def test_data_read():
     data_dir = Path(__file__).parent / "data"
     filename = data_dir / "Co_pos_0000.tpx3"
     # filename = data_dir / "Co_test_0000.tpx3"
-    stream = read(filename, dtype='<u8')
+    stream = read(filename)
 
     t_chunks = [ch for ch in chunks(stream)]
     assert len(t_chunks) == 178693
